change void ptr to templates; load and parse ships and containers

+++ 2_sem/metaheu_optim/exam_2/include/ship.hpp
@@ -14,7 +14,7 @@ struct Ship{
   std::string code;
   float cost;
   float max_weight;
-  float max_n_containers;
+  int max_n_containers;
 };
 
 #endif
